requirements.txt
grpcio
grpcio-tools
protobuf
pydantic
aiofiles
orjson
cryptography
python-dotenv